fix: change MILP time_limit arg hinted type from int to float

+++ optiwindnet/MILP/gurobi.py
@@ -39,7 +39,7 @@ class SolverGurobi(SolverPyomo, PoolHandler):
 
     def solve(
         self,
-        time_limit: int,
+        time_limit: float,
         mip_gap: float,
         options: dict[str, Any] = {},
         verbose: bool = False,
