+++ optiwindnet/MILP/ortools.py
@@ -89,7 +89,7 @@ class SolverORTools(Solver, PoolHandler):
 
     def solve(
         self,
-        time_limit: int,
+        time_limit: float,
         mip_gap: float,
         options: dict[str, Any] = {},
         verbose: bool = False,
