+++ optiwindnet/MILP/pyomo.py
@@ -108,7 +108,7 @@ class SolverPyomo(Solver):
 
     def solve(
         self,
-        time_limit: int,
+        time_limit: float,
         mip_gap: float,
         options: dict[str, Any] = {},
         verbose: bool = False,
